everything backend works

+++ tourney/models/round.py
@@ -62,56 +62,56 @@ class Round(models.Model):
     def clean(self):
         super().clean()
         errors = []
-        # if self.submit:
-        if self.p_team == self.d_team:
-            errors.append('one team cant compete against itself')
-        if self.p_team.next_side(self.pairing.round_num) == 'd':
-            errors.append(f"{self.p_team} is supposed to play d this round")
-        if self.d_team.next_side(self.pairing.round_num) == 'p':
-            errors.append(f"{self.d_team} is supposed to play p this round")
-        for round in self.p_team.p_rounds.all():
-            if round != self.instance and round.d_team == self.d_team:
-                errors.append(f"{self.p_team} and {self.d_team} played each other before")
-        for round in self.p_team.d_rounds.all():
-            if round != self.instance and round.p_team == self.d_team:
-                errors.append(f"{self.p_team} and {self.d_team} played each other before")
-        if self.presiding_judge == self.scoring_judge:
-            errors.append('assigning one judge for two roles')
+        if self.pairing.submit:
+            if self.p_team == self.d_team:
+                errors.append('one team cant compete against itself')
+            if self.p_team.next_side(self.pairing.round_num) == 'd':
+                errors.append(f"{self.p_team} is supposed to play d this round")
+            if self.d_team.next_side(self.pairing.round_num) == 'p':
+                errors.append(f"{self.d_team} is supposed to play p this round")
+            for round in self.p_team.p_rounds.all():
+                if round != self and round.d_team == self.d_team:
+                    errors.append(f"{self.p_team} and {self.d_team} played each other before")
+            for round in self.p_team.d_rounds.all():
+                if round != self and round.p_team == self.d_team:
+                    errors.append(f"{self.p_team} and {self.d_team} played each other before")
+            if self.presiding_judge == self.scoring_judge:
+                errors.append('assigning one judge for two roles')
 
-        if self.presiding_judge.preside == 0:
-            errors.append(f'{self.presiding_judge} can\'t preside')
+            if self.presiding_judge.preside == 0:
+                errors.append(f'{self.presiding_judge} can\'t preside')
 
-        for judge in self.judges:
-            if judge != None:
-                if not judge.get_availability(self.pairing.round_num):
-                    errors.append(f"{judge} is not available for Round {self.pairing.round_num}")
-                #check conflict
-                for team in self.teams:
-                    if team in judge.conflicts.all():
-                        errors.append(f"{judge} conflicted with p_team {team}")
-
-                #check if judged
-                judged = None
-                for round in judge.rounds:
-                    if round != self:
-                        if judged == None:
-                            judged = Team.objects.filter(pk=round.p_team.pk)
-                        else:
-                            judged |= Team.objects.filter(pk=round.p_team.pk)
-                        judged |= Team.objects.filter(pk=round.d_team.pk)
-                if judged != None:
+            for judge in self.judges:
+                if judge != None:
+                    if not judge.get_availability(self.pairing.round_num):
+                        errors.append(f"{judge} is not available for Round {self.pairing.round_num}")
+                    #check conflict
                     for team in self.teams:
-                        if team in judged:
-                            errors.append(f"{judge} has judged p_team {team}")
+                        if team in judge.conflicts.all():
+                            errors.append(f"{judge} conflicted with p_team {team}")
 
-                # #check if assigned in another division
-                pairings = Pairing.objects.filter(round_num=self.pairing.round_num)
-                if pairings.exists():
-                    for pairing in pairings.all():
-                        if pairing != self.pairing:
-                            for round in pairing.rounds.all():
-                                if judge in [round.presiding_judge,round.scoring_judge]:
-                                    errors.append(f"{judge} already assigned in {pairing.division}")
+                    #check if judged
+                    judged = None
+                    for round in judge.rounds:
+                        if round != self:
+                            if judged == None:
+                                judged = Team.objects.filter(pk=round.p_team.pk)
+                            else:
+                                judged |= Team.objects.filter(pk=round.p_team.pk)
+                            judged |= Team.objects.filter(pk=round.d_team.pk)
+                    if judged != None:
+                        for team in self.teams:
+                            if team in judged:
+                                errors.append(f"{judge} has judged p_team {team}")
+
+                    # #check if assigned in another division
+                    pairings = Pairing.objects.filter(round_num=self.pairing.round_num)
+                    if pairings.exists():
+                        for pairing in pairings.all():
+                            if pairing != self.pairing:
+                                for round in pairing.rounds.all():
+                                    if judge in [round.presiding_judge,round.scoring_judge]:
+                                        errors.append(f"{judge} already assigned in {pairing.division}")
 
         if errors != []:
             raise ValidationError(errors)
@@ -122,11 +122,12 @@ class Round(models.Model):
         super(Round, self).save()
         if is_new:
             CaptainsMeeting.objects.create(round=self)
-        if Ballot.objects.filter(round=self).exists():
-            Ballot.objects.filter(round=self).delete()
-        Ballot.objects.create(round=self, judge=self.presiding_judge)
-        Ballot.objects.create(round=self, judge=self.scoring_judge)
-        if self.extra_judge != None:
-            Ballot.objects.create(round=self, judge=self.extra_judge)
+        if self.submit:
+            if Ballot.objects.filter(round=self).exists():
+                Ballot.objects.filter(round=self).delete()
+            Ballot.objects.create(round=self, judge=self.presiding_judge)
+            Ballot.objects.create(round=self, judge=self.scoring_judge)
+            if self.extra_judge != None:
+                Ballot.objects.create(round=self, judge=self.extra_judge)
